[performance](exec) fix zone map in effective in column preidate (#58728)
### What problem does this PR solve?

When a page contains some NULL values, we set both the min value and its
associated flag to NULL for filtering purposes in IN and comparison
operations. In such cases, we can directly compare the min/max values to
determine whether the zone map is effective, without needing to handle
the scenario where the min value is NULL. However, if all values in the
page are NULL, the entire page can be directly filtered out.